Scrape historical tokens (to include currently inactive tokens)
